blog/requirements.txt
requests
boto3
flask-lambda-python36
simplejson